Refresh repository visibility to Netlify with empty commit
Private, organization-owned repositories is not free anymore with Netlify hosting.
While this repository is currently public, the last commit was made when it was private.
This means that Netlify does not have the updated settings.

A new commit must be made to the repository to trigger a deploy and send the updated settings to Netlify.

https://www.netlify.com/pricing/private-org-repo-faq/